fqdn: move FQDN API handlers to namemanager
Now that we've modularised the name manager, we can also move the
related cilium API handlers out of the daemon/cmd/fqdn.

This also allows slimming down the NameManager API further, hiding
internals.
