+++ daemon/cmd/api_handlers.go
@@ -37,10 +37,6 @@ type handlersOut struct {
 	EndpointPatchEndpointIDLabelsHandler endpoint.PatchEndpointIDLabelsHandler
 	EndpointPutEndpointIDHandler         endpoint.PutEndpointIDHandler
 
-	PolicyDeleteFqdnCacheHandler      policy.DeleteFqdnCacheHandler
-	PolicyGetFqdnCacheHandler         policy.GetFqdnCacheHandler
-	PolicyGetFqdnCacheIDHandler       policy.GetFqdnCacheIDHandler
-	PolicyGetFqdnNamesHandler         policy.GetFqdnNamesHandler
 	PolicyGetIdentityEndpointsHandler policy.GetIdentityEndpointsHandler
 	PolicyGetIdentityHandler          policy.GetIdentityHandler
 	PolicyGetIdentityIDHandler        policy.GetIdentityIDHandler
@@ -119,12 +115,6 @@ func ciliumAPIHandlers(dp promise.Promise[*Daemon], cfg *option.DaemonConfig, _ 
 
 	// /debuginfo
 	out.DaemonGetDebuginfoHandler = wrapAPIHandler(dp, getDebugInfoHandler)
-
-	// /fqdn/cache
-	out.PolicyGetFqdnCacheHandler = wrapAPIHandler(dp, getFqdnCacheHandler)
-	out.PolicyDeleteFqdnCacheHandler = wrapAPIHandler(dp, deleteFqdnCacheHandler)
-	out.PolicyGetFqdnCacheIDHandler = wrapAPIHandler(dp, getFqdnCacheIDHandler)
-	out.PolicyGetFqdnNamesHandler = wrapAPIHandler(dp, getFqdnNamesHandler)
 
 	// /ip/
 	out.PolicyGetIPHandler = wrapAPIHandler(dp, getIPHandler)
