+++ daemon/cmd/fqdn.go
@@ -11,15 +11,11 @@ import (
 	"strings"
 
 	"github.com/cilium/dns"
-	"github.com/go-openapi/runtime/middleware"
 	"github.com/sirupsen/logrus"
 
-	. "github.com/cilium/cilium/api/v1/server/restapi/policy"
-	"github.com/cilium/cilium/pkg/api"
 	"github.com/cilium/cilium/pkg/endpoint"
 	"github.com/cilium/cilium/pkg/fqdn"
 	"github.com/cilium/cilium/pkg/fqdn/dnsproxy"
-	"github.com/cilium/cilium/pkg/fqdn/matchpattern"
 	"github.com/cilium/cilium/pkg/fqdn/namemanager"
 	"github.com/cilium/cilium/pkg/fqdn/re"
 	"github.com/cilium/cilium/pkg/identity"
@@ -399,87 +395,4 @@ func (d *Daemon) notifyOnDNSMsg(
 	record.Log()
 
 	return nil
-}
-
-func getFqdnCacheHandler(d *Daemon, params GetFqdnCacheParams) middleware.Responder {
-	prefixMatcher, nameMatcher, source, err := parseFqdnFilters(params.Cidr, params.Matchpattern, params.Source)
-	if err != nil {
-		return api.Error(GetFqdnCacheBadRequestCode, err)
-	}
-
-	lookups, err := d.dnsNameManager.GetDNSHistoryModel("", prefixMatcher, nameMatcher, source)
-	switch {
-	case err != nil:
-		return api.Error(GetFqdnCacheBadRequestCode, err)
-	case len(lookups) == 0:
-		return NewGetFqdnCacheNotFound()
-	}
-
-	return NewGetFqdnCacheOK().WithPayload(lookups)
-}
-
-func deleteFqdnCacheHandler(d *Daemon, params DeleteFqdnCacheParams) middleware.Responder {
-	matchPatternStr := ""
-	if params.Matchpattern != nil {
-		matchPatternStr = *params.Matchpattern
-	}
-
-	err := d.dnsNameManager.DeleteDNSLookups(time.Now(), matchPatternStr)
-	if err != nil {
-		return api.Error(DeleteFqdnCacheBadRequestCode, err)
-	}
-	return NewDeleteFqdnCacheOK()
-}
-
-func getFqdnCacheIDHandler(d *Daemon, params GetFqdnCacheIDParams) middleware.Responder {
-	var epErr namemanager.NoEndpointIDMatch
-
-	prefixMatcher, nameMatcher, source, err := parseFqdnFilters(params.Cidr, params.Matchpattern, params.Source)
-	if err != nil {
-		return api.Error(GetFqdnCacheIDBadRequestCode, err)
-	}
-
-	lookups, err := d.dnsNameManager.GetDNSHistoryModel(params.ID, prefixMatcher, nameMatcher, source)
-	switch {
-	case errors.As(err, &epErr):
-		return api.Error(GetFqdnCacheIDNotFoundCode, err)
-	case err != nil:
-		return api.Error(GetFqdnCacheIDBadRequestCode, err)
-	case len(lookups) == 0:
-		return NewGetFqdnCacheIDNotFound()
-	}
-
-	return NewGetFqdnCacheIDOK().WithPayload(lookups)
-}
-
-func getFqdnNamesHandler(d *Daemon, params GetFqdnNamesParams) middleware.Responder {
-	payload := d.dnsNameManager.GetModel()
-	return NewGetFqdnNamesOK().WithPayload(payload)
-}
-
-func parseFqdnFilters(cidr, pattern, src *string) (fqdn.PrefixMatcherFunc, fqdn.NameMatcherFunc, string, error) {
-	prefixMatcher := func(ip netip.Addr) bool { return true }
-	if cidr != nil {
-		prefix, err := netip.ParsePrefix(*cidr)
-		if err != nil {
-			return nil, nil, "", err
-		}
-		prefixMatcher = func(ip netip.Addr) bool { return prefix.Contains(ip) }
-	}
-
-	nameMatcher := func(name string) bool { return true }
-	if pattern != nil {
-		matcher, err := matchpattern.ValidateWithoutCache(matchpattern.Sanitize(*pattern))
-		if err != nil {
-			return nil, nil, "", err
-		}
-		nameMatcher = func(name string) bool { return matcher.MatchString(name) }
-	}
-
-	source := ""
-	if src != nil {
-		source = *src
-	}
-
-	return prefixMatcher, nameMatcher, source, nil
 }
