+++ pkg/fqdn/namemanager/manager.go
@@ -12,12 +12,10 @@ import (
 	"regexp"
 	"slices"
 
-	"github.com/go-openapi/strfmt"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/sync/errgroup"
 	"k8s.io/apimachinery/pkg/util/sets"
 
-	"github.com/cilium/cilium/api/v1/models"
 	"github.com/cilium/cilium/pkg/controller"
 	"github.com/cilium/cilium/pkg/endpoint"
 	"github.com/cilium/cilium/pkg/fqdn"
@@ -37,19 +35,6 @@ import (
 	"github.com/cilium/cilium/pkg/time"
 )
 
-var (
-	DNSSourceLookup     = "lookup"
-	DNSSourceConnection = "connection"
-)
-
-type NoEndpointIDMatch struct {
-	ID string
-}
-
-func (e NoEndpointIDMatch) Error() string {
-	return "unable to find target endpoint ID: " + e.ID
-}
-
 // The implementation of the NameManager interface.
 type manager struct {
 	lock.RWMutex
@@ -62,7 +47,6 @@ type manager struct {
 	// use these selectors to map selectors --> IPs.
 	allSelectors map[api.FQDNSelector]*regexp.Regexp
 
-	// cache is a private copy of the pointer from config.
 	cache *fqdn.DNSCache
 
 	bootstrapCompleted bool
@@ -100,101 +84,6 @@ func New(params ManagerParams) *manager {
 	}
 
 	return n
-}
-
-// GetModel returns the API model of the NameManager.
-func (n *manager) GetModel() *models.NameManager {
-	n.RWMutex.RLock()
-	defer n.RWMutex.RUnlock()
-
-	allSelectors := make([]*models.SelectorEntry, 0, len(n.allSelectors))
-	for fqdnSel, regex := range n.allSelectors {
-		pair := &models.SelectorEntry{
-			SelectorString: fqdnSel.String(),
-			RegexString:    regex.String(),
-		}
-		allSelectors = append(allSelectors, pair)
-	}
-
-	return &models.NameManager{
-		FQDNPolicySelectors: allSelectors,
-	}
-}
-
-// GetDNSHistoryModel returns API models.DNSLookup copies of DNS data in each
-// endpoint's DNSHistory. These are filtered by the specified matchers if
-// they are non-empty.
-//
-// Note that this does *NOT* dump the NameManager's own global DNSCache.
-//
-// endpointID may be "" in order to get DNS history for all endpoints.
-func (n *manager) GetDNSHistoryModel(endpointID string, prefixMatcher fqdn.PrefixMatcherFunc, nameMatcher fqdn.NameMatcherFunc, source string) (lookups []*models.DNSLookup, err error) {
-	eps := n.getEndpointsDNSInfo(endpointID)
-	if eps == nil {
-		return nil, &NoEndpointIDMatch{ID: endpointID}
-	}
-	for _, ep := range eps {
-		lookupSourceEntries := []*models.DNSLookup{}
-		connectionSourceEntries := []*models.DNSLookup{}
-		for _, lookup := range ep.DNSHistory.Dump() {
-			if !nameMatcher(lookup.Name) {
-				continue
-			}
-
-			// The API model needs strings
-			IPStrings := make([]string, 0, len(lookup.IPs))
-
-			// only proceed if any IP matches the prefix selector
-			anIPMatches := false
-			for _, ip := range lookup.IPs {
-				anIPMatches = anIPMatches || prefixMatcher(ip)
-				IPStrings = append(IPStrings, ip.String())
-			}
-			if !anIPMatches {
-				continue
-			}
-
-			lookupSourceEntries = append(lookupSourceEntries, &models.DNSLookup{
-				Fqdn:           lookup.Name,
-				Ips:            IPStrings,
-				LookupTime:     strfmt.DateTime(lookup.LookupTime),
-				TTL:            int64(lookup.TTL),
-				ExpirationTime: strfmt.DateTime(lookup.ExpirationTime),
-				EndpointID:     int64(ep.ID64),
-				Source:         DNSSourceLookup,
-			})
-		}
-
-		for _, delete := range ep.DNSZombies.DumpAlive(prefixMatcher) {
-			for _, name := range delete.Names {
-				if !nameMatcher(name) {
-					continue
-				}
-
-				connectionSourceEntries = append(connectionSourceEntries, &models.DNSLookup{
-					Fqdn:           name,
-					Ips:            []string{delete.IP.String()},
-					LookupTime:     strfmt.DateTime(delete.AliveAt),
-					TTL:            0,
-					ExpirationTime: strfmt.DateTime(ep.DNSZombies.NextCTGCUpdate),
-					EndpointID:     int64(ep.ID64),
-					Source:         DNSSourceConnection,
-				})
-			}
-		}
-
-		switch source {
-		case DNSSourceLookup:
-			lookups = append(lookups, lookupSourceEntries...)
-		case DNSSourceConnection:
-			lookups = append(lookups, connectionSourceEntries...)
-		default:
-			lookups = append(lookups, lookupSourceEntries...)
-			lookups = append(lookups, connectionSourceEntries...)
-		}
-	}
-
-	return lookups, nil
 }
 
 // getEndpointsDNSInfo is used by the NameManager to iterate through endpoints
